requirements.txt
django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
django-filter==23.3
requests==2.32.4